Disabled setting the refactoring flag
The delete operation is also a refactoring.
The replayed uses a refactoring flag to suppress replaying refactoring
operations.
Therefore the delete was not performed, since it was during its own
Delete refactoring.

I removed the flag being set. It is not a pretty solution, but now I
cannot foresee how we’ll reuse the refactoring replay architecture so I
just went with what was easier.