Fixed bug that prevented replaying at high speed.
The bug was caused by a previous task in which the currentEditor field was removed. That task required the removal of an if statement that tested wether the currentEditor was equal to an expected editor. If they weren't equal, the execution thread got interrupted. Without thaat if, the code that interrupted the execution thread got always called.

I also commented out the contents of the if to remove the bug.